fix(event): modificación función insert y update

+++ app/crud/route.py
@@ -1,3 +1,5 @@
+from fastapi import HTTPException
+
 from sqlalchemy.orm import Session
 from app.models.domain.route import Route
 from app.models.schema.route import RouteCreate, RouteUpdate
@@ -8,9 +10,6 @@ def create_route(db: Session, route_data: RouteCreate):
     db.commit()
     db.refresh(new_route)
     return new_route
-
-def get_route_by_id(db: Session, route_id: int):
-    return db.query(Route).filter(Route.id == route_id).first()
 
 def get_routes(db: Session, skip: int = 0, limit: int = 10):
     return db.query(Route).offset(skip).limit(limit).all()
@@ -28,7 +27,14 @@ def update_route(db: Session, route_id: int, route_data: RouteUpdate):
 def delete_route(db: Session, route_id: int):
     route = db.query(Route).filter(Route.id == route_id).first()
     if not route:
-        return None
+        raise HTTPException(status_code=404, detail="Ruta no encontrada")
+
+    if route.events:  # ⚠️ Verifica si tiene eventos asociados
+        raise HTTPException(
+            status_code=400,
+            detail="No se puede eliminar la ruta porque está asociada a uno o más eventos."
+        )
+
     db.delete(route)
     db.commit()
     return route
